feat: move copytree_exists_ok to base
As suggested by @fkraushofer
